Removed global variable calls, cleaned up static.

+++ storage/myisam/myisam.h
@@ -248,11 +248,11 @@ typedef struct st_columndef		/* column information */
 
 
 extern char * myisam_log_filename;		/* Name of logfile */
-extern ulong myisam_block_size;
-extern ulong myisam_concurrent_insert;
+extern uint32_t myisam_block_size;
+extern uint32_t myisam_concurrent_insert;
 extern bool myisam_flush,myisam_delay_key_write,myisam_single_user;
 extern my_off_t myisam_max_temp_length;
-extern ulong myisam_bulk_insert_tree_size, myisam_data_pointer_size;
+extern uint32_t myisam_bulk_insert_tree_size, myisam_data_pointer_size;
 
 	/* Prototypes for myisam-functions */
